feat(transactions): add transaction form, validation and API layer
- Implements the core transaction management feature using a feature-based architecture.
- Adds a fully validated form with React Hook Form and Zod, integrated with HeroUI components and TanStack Query hooks for predictable API state management.
- Introduces shared utilities for UTC-safe date handling and currency formatting, along with updates global styles.

+++ src/shared/utils/parseEnvNumber.ts
@@ -2,6 +2,6 @@ export function parseEnvNumber(
   value: string | undefined,
   fallback: number,
 ): number {
-  const parsed = Number.parseInt(value || "");
+  const parsed = Number.parseFloat(value || "");
   return Number.isNaN(parsed) ? fallback : parsed;
 }
